+++ CudaBEM/Point.h
@@ -4,9 +4,9 @@
 
 struct Point
 {
-	double x, y, z;
+	float x, y, z;
 
-	__host__ __device__  Point(double x, double y, double z) : x(x), y(y), z(z) {}
+	__host__ __device__  Point(float x, float y, float z) : x(x), y(y), z(z) {}
 	__host__ __device__  Point() : x(0.0), y(0.0), z(0.0) {}
 
 	__host__ __device__ Point operator+(const Point& a)
@@ -19,22 +19,22 @@ struct Point
 		return Point(x - a.x, y - a.y, z - a.z);
 	}
 
-	__host__ __device__ double operator*(const Point& a)
+	__host__ __device__ float operator*(const Point& a)
 	{
 		return x * a.x + y * a.y + z * a.z;
 	}
 
-	__host__ __device__ Point operator*(double constant)
+	__host__ __device__ Point operator*(float constant)
 	{
 		return Point(x * constant, y * constant, z * constant);
 	}
 
-	__host__ __device__ Point operator/(double constant)
+	__host__ __device__ Point operator/(float constant)
 	{
 		return Point(x / constant, y / constant, z / constant);
 	}
 
-	__host__ __device__ double Norm()
+	__host__ __device__ float Norm()
 	{
 		return sqrt(Point(x, y, z) * Point(x, y, z));
 	}
